Added class Product Divided classes into files Calculator class modification and rename file and class Fix class Converter
