+++ src/main/java/Main.java
@@ -1,8 +1,8 @@
 
 public class Main {
     public static void main(String[] args) {
-        Calculater calculater = new Calculater();
-        calculater.countingProducts();
+        Calculator calculator = new Calculator();
+        calculator.countingProducts();
     }
 }
