+++ src/main/java/Product.java
@@ -1,0 +1,9 @@
+public class Product {
+    String name;
+    float price;
+
+    Product(String name, float price){
+        this.name = name;
+        this.price = price;
+    }
+}
